libretro: Move the core's game config save folder.
Previously it was located alongside the saves, screenshots and
other folders in retroarch's root as "savesfreej2me". With thi
change, it will now be moved into the saves folder, and be named
as "freej2meconfig", which makes more sense.

However, that's just for the config folder. The "rms" one that
contains the actual game saves is still created and accessed at
the root of the folder that the frontent binary was called from,
so if you symlinked an AppImage of it to your desktop, the core
will create a "rms" folder there to put the game saves and won't
be able to access it when you open the AppImage directly, where it
will create a new rms folder alongside the .AppImage.

But maybe that's only happening on my linux machine through an
AppImage Retroarch build, more testing is needed.