libretro: Prepare and organize the to-be-exposed core options
This commit is a cleaner approach to the multiple commits on my
master-dev branch aimed at reorganizing how the libretro core's
configs are presented. Instead of having a lot of commits and
reverts on small tidbits here and there, this single commit
puts all of the to-be-exposed configs in a separate file
"freej2me_libretro.h", and the main source code only has to
include that header in order to access them. All configs also
have variants for each different core_options_version within
Libretro, allowing it to work well on different frontends.