libretro + config: Clean up, and allow core settings to be tweaked
This commit should allow the exposed core configs to be tweaked
through the frontend without relying on FreeJ2ME's built-in menu.
A string containing the settings is passed through a pipe from
the core to FreeJ2ME's java app, where the string is tokenized and
each individual setting is parsed and updated in the config file
without the need for a core restart. A fix for cases where new
games didn't have their own .conf files yet, or had them deleted
didn't get configs matching the frontend's setup is also in place,
working as init arguments passed to Java through the commandline
at boot.