PlatformPlayer: Implement initial IMA ADPCM decoding support
In order to decode IMA ADPCM wav files, we first read the
incoming wav's header and check if the reported Audio Format is
'17', indicating that it is an IMA ADPCM stream. If it is, we
begin to decode its samples into signed PCM16LE while maintaining
the same amount of channels, and sampling rate. Once that is done,
we build a new header for the decoded stream to reflect its new
format and then send the completed stream back to PlatformPlayer's
wavPlayer so it can handle it like a standard wav stream.

Additionally, a simple 'low-pass filter' was added to combat
crackling on resulting samples, as it was really overbearing, it
vastly improves quality at almost no cost.