Config: FreeJ2ME: Libretro: Support HW Acceleration for Java2D API
FreeJ2ME is rather CPU-heavy since it relies on CPU for rendering,
audio, and pretty much everything else it needs to do in order to
run J2ME apps. This commit allows it to offload everything that
uses the Java2D API to the Graphics Adapter, making the graphics
pipeline much faster and lowering cpu overhead in more complex
scenes.