Anbu: Introduce full mouse support for the SDL2 UI.
After playing around with AWT for a bit, i came to notice some
quite serious shortcomings for an app like FreeJ2ME. Mainly,
supporting joypads and other mapping devices would rely on Java
libs, which are  few and far between, and would bloat up the main
jar far too much for my liking. So despite finding it much easier
to work with as far as UI goes, i think SDL2 is what we should
be striving to use for the standalone/non-libretro flavor of
FreeJ2ME.

But enough rambling about the future, onto the commit:

This introduces full mouse support into Anbu/SDL2, dragging
included. It should be on par with AWT and libretro as far as
mouse input goes. Although this brought up a few limitations i
also intend to fix shortly:

1 - Scaling in general of the SDL2 window wasn't very good and it
was easy to end up with a window that had a different aspect ratio
compared to the frames FreeJ2ME was sending into it. For now, the
window is forced to the same resolution as the source material,
that is, the supposed "jar" resolution, although i intend to
introduce window scaling through the "+ -" key combo.

2 - Fullscreen is a problem of varying degrees depending on the
user's monitor and the app being ran in FreeJ2ME. Mouse coordinates
are polled in regards to the window size and not the frame's... it
should be possible to make those coordinates window-independent and
only rely on the frame size, but given how much those can change
depending on app res and rotation, i don't know how feasible it would
be yet. Might be easier to just force the fullscreen rect to be the
same aspect ratio from the frame as well.

3 - We might not have a use for FreeImage anymore, since we can
take screenshots by using SDL_Image, which is already part of the
SDL2 package... and since the frame should be getting the whole
window size in order not to mess with pointer coordinates, we won't
have a use for BG images either. I see it as a plus, as the less
external deps we have here, the better.