Libretro: Allow D-Pad/Analog Stick to be used for the whole keypad
Helps certain games like "Time Crisis Elite" which dedicate each
keypad's key to a specific area of the screen and etc. But also
impacts other games negatively, which is why this is a toggle.

And while on the topic of Time Crisis Elite, this one used ONLY
the arrow keys for menu navigation, and ONLY the keypad during
gameplay, so key with "overlapping" functionality such as 2/Up,
4/Left are now both sent to the emulated phone, we don't have
enough buttons on most joypads to physically accommodate for both...
and it would also be a mess if we could, probably.